[dagstermill] fix test imports
