+++ examples/docs_snippets/tox.ini
@@ -3,12 +3,16 @@ skipsdist = true
 
 [testenv]
 download = True
+setenv =
+    VIRTUALENV_SETUPTOOLS=77.0.3
 passenv =
     CI_*
     COVERALLS_REPO_TOKEN
     POSTGRES_TEST_DB_HOST
     BUILDKITE*
-install_command = uv pip install -b ../../python_modules/libraries/dagster-pyspark/build-constraints {opts} {packages}
+; dagstermill test errors prevent using uv, require compat mode install
+; install_command = uv pip install {opts} {packages}
+install_command = python -m pip install {opts} {packages} --config-settings editable_mode=compat
 deps =
   -e ../../python_modules/dagster[test]
   -e ../../python_modules/dagster-pipes
@@ -48,7 +52,6 @@ deps =
   -e .[full]
 allowlist_externals =
   /bin/bash
-  uv
 commands =
   !windows: /bin/bash -c '! pip list --exclude-editable | grep -e dagster'
   pytest -c ../../pyproject.toml -vv {posargs}
