+++ python_modules/libraries/dagstermill/dagstermill_tests/test_ops.py
@@ -2,13 +2,11 @@ import importlib.util
 import os
 import pickle
 import tempfile
-from contextlib import contextmanager
 
 import nbformat
 import pytest
 from dagster import execute_job, job
 from dagster._check import CheckError
-from dagster._core.definitions.metadata import NotebookMetadataValue, PathMetadataValue
 from dagster._core.definitions.reconstruct import ReconstructableJob
 from dagster._core.storage.tags import COMPUTE_KIND_TAG
 from dagster._core.test_utils import instance_for_test
@@ -17,50 +15,13 @@ from dagstermill import DagstermillError
 from dagstermill.compat import ExecutionError
 from dagstermill.examples.repository import custom_io_mgr_key_job
 from dagstermill.factory import define_dagstermill_op
+from dagstermill.test_utils import cleanup_result_notebook, exec_for_test, get_path
 from jupyter_client.kernelspec import NoSuchKernel
 from nbconvert.preprocessors import ExecutePreprocessor
 
 DAGSTER_PANDAS_PRESENT = importlib.util.find_spec("dagster_pandas") is not None
 SKLEARN_PRESENT = importlib.util.find_spec("sklearn") is not None
 MATPLOTLIB_PRESENT = importlib.util.find_spec("matplotlib") is not None
-
-
-def get_path(materialization_event):
-    for value in materialization_event.event_specific_data.materialization.metadata.values():
-        if isinstance(value, (PathMetadataValue, NotebookMetadataValue)):
-            return value.path
-
-
-def cleanup_result_notebook(result):
-    if not result:
-        return
-    materialization_events = [
-        x for x in result.all_events if x.event_type_value == "ASSET_MATERIALIZATION"
-    ]
-    for materialization_event in materialization_events:
-        result_path = get_path(materialization_event)
-        if os.path.exists(result_path):  # pyright: ignore[reportArgumentType]
-            os.unlink(result_path)  # pyright: ignore[reportArgumentType]
-
-
-@contextmanager
-def exec_for_test(fn_name, env=None, raise_on_error=True, **kwargs):
-    result = None
-    recon_job = ReconstructableJob.for_module("dagstermill.examples.repository", fn_name)
-
-    with instance_for_test() as instance:
-        try:
-            with execute_job(
-                job=recon_job,
-                run_config=env,
-                instance=instance,
-                raise_on_error=raise_on_error,
-                **kwargs,
-            ) as result:
-                yield result
-        finally:
-            if result:
-                cleanup_result_notebook(result)
 
 
 @pytest.mark.notebook_test
